Project checkout commit fuzzy
In order to checkout a certain commit of the project's git
directory, try to checkout every subdirectory of $SRC. We do
this, since we don't know for sure how the directory the project
is cloned to will be named. Might be the project's name, might
be something different (e. g. project name: freetype2, project
directory: freetype2-testing).